% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reference_funcs.R
\name{gdl_countries}
\alias{gdl_countries}
\title{Get country list}
\usage{
gdl_countries(session)
}
\arguments{
\item{session}{A valid GDL session object to interface with.}
}
\value{
A data frame containing a list of countries for the dataset.
}
\description{
Returns a list of countries available in the current dataset.
}
\examples{
\dontrun{
# Create a session using your API token (provided by environment here)
session <- gdl_session(Sys.getenv('GDL_API_TOKEN'))
# Request list of countries
countries <- gdl_countries(session)
head(countries, n=10)
}
}
