% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reference_funcs.R
\name{gdl_datasets}
\alias{gdl_datasets}
\title{Get dataset list}
\usage{
gdl_datasets(session)
}
\arguments{
\item{session}{A valid GDL session object to interface with.}
}
\value{
A data frame containing a list of datasets.
}
\description{
Returns a list of datasets available.
}
\examples{
\dontrun{
# Create a session using your API token (provided by environment here)
session <- gdl_session(Sys.getenv('GDL_API_TOKEN'))
# Request list of datasets
datasets <- gdl_datasets(session)
head(datasets, n=10)
}
}
