% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdl_session.R
\name{gdl_session}
\alias{gdl_session}
\title{GDL session constructor}
\usage{
gdl_session(token)
}
\arguments{
\item{token}{A valid GDL API token, obtainable from GlobalDataLab.org}
}
\value{
A GDL session object for the token with default indicators set.
}
\description{
Returns a new GDL session object
}
\examples{
\dontrun{
# Create a session using your API token (provided by environment here)
session <- gdl_session(Sys.getenv('GDL_API_TOKEN'))
}
}
