% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setter_funcs.R
\name{set_countries_all}
\alias{set_countries_all}
\title{Set session to retrieve data for all available countries}
\usage{
set_countries_all(session)
}
\arguments{
\item{session}{A valid GDL session object to interface with.}
}
\value{
An amended GDL session object.
}
\description{
Switch the session to retrieve data for all available countries,
rather than a specific set of countries.
}
\examples{
\dontrun{
# Create a session using your API token (provided by environment here)
session <- gdl_session(Sys.getenv('GDL_API_TOKEN'))
session <- set_countries_all(session)
}
}
