% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setter_funcs.R
\name{set_country}
\alias{set_country}
\title{Set country to retrieve data for}
\usage{
set_country(session, country)
}
\arguments{
\item{session}{A valid GDL session object to interface with.}

\item{country}{An ISO3 country code.}
}
\value{
An amended GDL session object.
}
\description{
The country to retrieve GDL indicator data for.
}
\examples{
\dontrun{
# Create a session using your API token (provided by environment here)
session <- gdl_session(Sys.getenv('GDL_API_TOKEN'))
session <- set_country(session, 'IND')
}
}
