% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setter_funcs.R
\name{set_dataset}
\alias{set_dataset}
\title{Set session to retrieve data from a particular dataset}
\usage{
set_dataset(session, dataset)
}
\arguments{
\item{session}{A valid GDL session object to interface with.}

\item{dataset}{Dataset identifier (string)}
}
\value{
An amended GDL session object.
}
\description{
Switch the session to retrieve data from the dataset specified.
}
\examples{
\dontrun{
# Create a session using your API token (provided by environment here)
session <- gdl_session(Sys.getenv('GDL_API_TOKEN'))
session <- set_dataset(session, 'shdi')
}
}
