% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setter_funcs.R
\name{set_interpolation}
\alias{set_interpolation}
\title{Set interpolation state}
\usage{
set_interpolation(session, state)
}
\arguments{
\item{session}{A valid GDL session object to interface with.}

\item{state}{Whether or not to use interpolation (boolean)}
}
\value{
An amended GDL session object.
}
\description{
Turns interpolation on (T) or off (F).
}
\examples{
\dontrun{
# Create a session using your API token (provided by environment here)
session <- gdl_session(Sys.getenv('GDL_API_TOKEN'))
session <- set_interpolation(session, TRUE)
}
}
