% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setter_funcs.R
\name{set_levels}
\alias{set_levels}
\title{Set data levels to retrieve data for}
\usage{
set_levels(session, levels)
}
\arguments{
\item{session}{A valid GDL session object to interface with.}

\item{levels}{Vector of level identifiers (integers)}
}
\value{
An amended GDL session object.
}
\description{
Specify which data levels to retrieve data for.
A list of levels may be obtained through \code{gdl_levels}.
}
\examples{
\dontrun{
# Create a session using your API token (provided by environment here)
session <- gdl_session(Sys.getenv('GDL_API_TOKEN'))
session <- set_levels(session, c(1,4))
}
}
