% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setter_funcs.R
\name{set_transposed}
\alias{set_transposed}
\title{Enable transposition of years to variables.}
\usage{
set_transposed(session, value)
}
\arguments{
\item{session}{A valid GDL session object to interface with.}

\item{value}{Turn transposition on (TRUE) or off (FALSE).}
}
\value{
An amended GDL session object.
}
\description{
This enables transposition of years to variables. Note that this only applies
to single-variable queries.
}
\examples{
\dontrun{
# Create a session using your API token (provided by environment here)
session <- gdl_session(Sys.getenv('GDL_API_TOKEN'))
session <- set_transposed(session, TRUE)
}
}
