% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setter_funcs.R
\name{set_year}
\alias{set_year}
\title{Set year to retrieve data for}
\usage{
set_year(session, year)
}
\arguments{
\item{session}{A valid GDL session object to interface with.}

\item{year}{Year to retrieve data for (integer)}
}
\value{
An amended GDL session object.
}
\description{
Specify which year to retrieve data for.
Only used when retrieving multiple indicators at once.
}
\examples{
\dontrun{
# Create a session using your API token (provided by environment here)
session <- gdl_session(Sys.getenv('GDL_API_TOKEN'))
session <- set_year(session, 2021)
}
}
