% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdl_session.R
\docType{data}
\name{show.GDLSession}
\alias{show.GDLSession}
\title{GDLSession show function}
\format{
An object of class \code{character} of length 1.
}
\usage{
show.GDLSession
}
\description{
This is a user-friendly show function for the GDLSession class,
hiding internals from simple print statements.
}
\keyword{datasets}
